Large VueJS Rework
VueJS is now the main framework of the project.
Just about all functionality that the site used to have is remade.
Booking calendar is the current problem being handled.
Any looking back at old versions of code can be seen hosted on the new paltz site or at the local copy being held on my laptop.