2/23 Update
setInterval() loops have been replaced with recursive setTimeout() to stop buildups.

New time highlighter feature has been implemented so that the 15 window that the user is currently in is highlighted to speed up the time needed for a user to find the slot they want.

The next step is to finally start creating the weekly view calendar. After being told the answer and feeling stupid I've realized that the week range to be shown should be Sunday to Saturday - Thanks Caitlin.