2/14 Update
Weekly view is almost complete:
Most server side booking fetching is in place with a possible problem due to mySQL being based in a different time zone.
CSS is the main next step with the new calender.

Time window highlighter was updated to update the old highlighted windows after the time moved to a new hour.