4/23 Update
Changes to BookedModal to enable deleation of bookings with proper permissions.
This is only checked on the front end so a backend portion will have to be made.
This should be enough for the presentation though.

Header navigation has been reamped and lets the user know what page they are on.
Same affect should be added to the pages with multiple views to them.

Anylasis page has some nice custom made mulicheck boxes.